Added user authorisation to journal entry access,update and delete

+++ src/main/java/com/edigest/journalApp/controller/JourneyEntryController.java
@@ -14,6 +14,7 @@ import org.springframework.web.bind.annotation.*;
 
 import java.time.LocalDateTime;
 import java.util.*;
+import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping("/journal")
@@ -60,38 +61,71 @@ public class JourneyEntryController {
     // Request Parameter is when input is with ? in url/Params eg journal/getEntry/id?id=2
     @GetMapping("/getEntry/id/{myId}")
     public ResponseEntity<JourneyEntry> getEntry(@PathVariable ObjectId myId){
-        try{
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String username = authentication.getName(); //  we get user trying to view myID journal entry
+        // we fetch all journal entries corresponding this username and filter in it to find journal entry with myId
+        List<JourneyEntry> journeyEntryList = userService.findUserByUsername(username).getJourneyEntries().stream()
+                .filter(x -> x.getId().equals(myId)).collect(Collectors.toList());
+
+        if(!journeyEntryList.isEmpty()){
             Optional<JourneyEntry> obj = journalEntryService.getEntryById(myId);
             if(obj.isPresent()){
                 return new ResponseEntity<>(obj.get(),HttpStatus.OK);
             }else{
                 return new ResponseEntity<>(HttpStatus.NOT_FOUND);
             }
-        }catch (Exception e){
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
+        return new ResponseEntity<>(HttpStatus.FORBIDDEN); // 403 authenticated not authorised
+
 
     }
 
-    @PutMapping("/editEntry/{username}/{myId}")
-    public ResponseEntity<JourneyEntry> updateJournalEntry(@PathVariable ObjectId myId,String username,@RequestBody JourneyEntry req){
+    @PutMapping("/editEntry/{myId}")
+    public ResponseEntity<JourneyEntry> updateJournalEntry(@PathVariable ObjectId myId,@RequestBody JourneyEntry req){
         try{
-            journalEntryService.updateEntry(myId,username,req);
-            return new ResponseEntity<>(HttpStatus.OK);
+            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+            String username = authentication.getName();
+            // to check if this user has corresponding myId journal entry or not
+            List<JourneyEntry> journeyEntryList = userService.findUserByUsername(username).getJourneyEntries().stream()
+                            .filter(x -> x.getId().equals(myId)).toList();
+
+            // if myID journal entry is present in User of username then they can edit it
+            if(!journeyEntryList.isEmpty()){
+                journalEntryService.updateEntry(myId,username,req);
+                return new ResponseEntity<>(HttpStatus.OK);
+            }
+            else return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+
         }catch (Exception e){
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
 
     }
 
-    @DeleteMapping("/deleteEntry/{username}/{myId}")
-    public ResponseEntity<?> deleteEntry(@PathVariable ObjectId myId,@PathVariable String username){
-        try {
-            journalEntryService.deleteEntry(myId,username);
-            return new ResponseEntity<>(HttpStatus.OK);
-        }catch (Exception e) {
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+    @DeleteMapping("/deleteEntry/{myId}")
+    public ResponseEntity<?> deleteEntry(@PathVariable ObjectId myId){
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        String username = auth.getName();
+
+        List<JourneyEntry> entryList = userService.findUserByUsername(username).getJourneyEntries().stream()
+                .filter(x -> x.getId().equals(myId)).toList();
+
+        if(!entryList.isEmpty()){
+            try {
+                boolean ifRemoved = journalEntryService.deleteEntry(myId,username);
+                if(ifRemoved){
+                    return new ResponseEntity<>(HttpStatus.OK);
+                }else{
+                    return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+                }
+
+            }catch (Exception e) {
+                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            }
+        }else{
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN); // 403 authenticated not authorised
         }
+
     }
 
 }
